feat: auto-update, WASM plugin sandbox, perf validation, native polish, CI fixes
Auto-update (zeroterm):
- --version/-V prints 'zeroterm 0.2.0' (env CARGO_PKG_VERSION, no hardcode)
- upgrade subcommand reuses install.sh upgrade (spawns it) + curl one-liner fallback
- Fix hardcoded 'v0.1.0' log; install.sh macos asset naming aligned (macos-<arch>.zip)
- docs/packaging.md: remove version-flag TODO, document --version/upgrade

WASM plugin sandbox (Phase 5) — new crate craters/zeroterm-plugin:
- PluginHost/Plugin via wasmtime 28 + wasmtime-wasi (preview1), WASI stdin/stdout ABI
- StoreLimits memory cap (16MiB) + fuel (infinite-loop kill), proc_exit detection
- 5 tests (echo, fuel-kill, memory-kill, proc_exit 0/3); workspace 173 pass

Perf validation (docs/perf.md + 3 new benches + scripts/bench.sh):
- 8 criterion benches; parser NOT bottleneck (1.4-5.4x past 120fps@4K 7.78M cells/s)
- renderer uses Backends::PRIMARY (Vulkan+Metal+DX12 auto) — no change needed

Native polish:
- make_macos_app.sh: arm64 naming, hardened-runtime codesign, real notarytool
  notarize()+staple+re-zip (env-driven, skips w/o creds)
- docs/platforms.md: macOS Metal/notarization, Windows ConPTY/DX12/WiX/signtool

CI robustness (was failing every run):
- Fix cargo fmt --check (6 files) + 10+ clippy -D warnings violations
- ci.yml: fail-fast:false, 60min timeout, --locked, dtolnay+Swatinem, Windows
  job excludes zeroterm+zeroterm-ssh (main.rs unportable — documented)
- release.yml: ubuntu+macos build matrix, --locked, missing flatpak stubs added

+++ crates/zeroterm-core/src/cell.rs
@@ -122,9 +122,9 @@ impl Color {
 
     pub fn from_ansi_216(idx: u8) -> Color {
         let idx = idx - 16;
-        let r = (idx / 36) as u8;
-        let g = ((idx % 36) / 6) as u8;
-        let b = (idx % 6) as u8;
+        let r = idx / 36;
+        let g = (idx % 36) / 6;
+        let b = idx % 6;
         let map = |v: u8| if v == 0 { 0 } else { v * 40 + 55 };
         Color {
             r: map(r),
